pyproject.toml
[tool.poetry]
name = "waffle-window"
version = "0.1.0"
description = "Django app for managing user groups and django-waffle flags."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.md"
homepage = "https://github.com/yunojuno/django-waffle-window"
repository = "https://github.com/yunojuno/django-waffle-window"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2 || ^3.0"
django-waffle = "^0.19.0"

[tool.poetry.dev-dependencies]
tox = "^3.14.3"
pytest = "^3.0"
pytest-cov = "^2.8"
pytest-django = "^3.7"
coverage = "^4.5"
pylint = "^2.4"
pylint-django = "^2.0"
flake8 = "^3.7"
flake8-bandit = "^2.1"
flake8-docstrings = "^1.5"
isort = "^4.3"
mypy = "^0.750.0"
pre-commit = "^1.20"
black = "=19.3b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
